chore(server): remove view group v1
