+++ packages/twenty-server/src/engine/metadata-modules/view-group/view-group.module.ts
@@ -8,7 +8,6 @@ import { PermissionsModule } from 'src/engine/metadata-modules/permissions/permi
 import { ViewGroupController } from 'src/engine/metadata-modules/view-group/controllers/view-group.controller';
 import { ViewGroupEntity } from 'src/engine/metadata-modules/view-group/entities/view-group.entity';
 import { ViewGroupResolver } from 'src/engine/metadata-modules/view-group/resolvers/view-group.resolver';
-import { ViewGroupV2Service } from 'src/engine/metadata-modules/view-group/services/view-group-v2.service';
 import { ViewGroupService } from 'src/engine/metadata-modules/view-group/services/view-group.service';
 import { ViewPermissionsModule } from 'src/engine/metadata-modules/view-permissions/view-permissions.module';
 import { ViewEntity } from 'src/engine/metadata-modules/view/entities/view.entity';
@@ -27,7 +26,7 @@ import { WorkspaceMigrationV2Module } from 'src/engine/workspace-manager/workspa
     ViewPermissionsModule,
   ],
   controllers: [ViewGroupController],
-  providers: [ViewGroupService, ViewGroupV2Service, ViewGroupResolver],
+  providers: [ViewGroupService, ViewGroupResolver],
   exports: [ViewGroupService],
 })
 export class ViewGroupModule {}
